feat(governor): wire GroupPolicyStore into ProcessGovernor, add deterministic eviction tests
Wave2 P2-2: wire GroupPolicyStore into ProcessGovernor with:
- GroupPolicyStore member (cap=256) and pid→group map (cap=8192)
- LRU eviction for groups and pids
- apply_group_policy and apply_cgroup_policy methods
- Event callbacks for GROUP_EVICT/PIDMAP_EVICT/GOV_CGROUP_UNAVAILABLE
- Rate-limited cgroup unavailability events
- ViolationAction aligned to NONE/WARN/SOFT_KILL/HARD_KILL
- Deterministic eviction unit tests using controllable time sequence
- Fix memset-on-non-trivial warnings in GroupPolicyStore
- Add GovEventType enum registry to prevent event ID collisions

+++ include/heidi-kernel/gov_rule.h
@@ -4,6 +4,7 @@
 #include <cstdint>
 #include <optional>
 #include <string>
+#include <string_view>
 
 namespace heidi {
 namespace gov {
@@ -42,7 +43,7 @@ struct CpuPolicy {
   std::optional<int8_t> nice;
   std::optional<uint8_t> max_pct;
   std::optional<uint32_t> period_us;
-  std::optional<uint32_t> quota_us;
+  std::optional<uint64_t> quota_us;
 };
 
 struct MemPolicy {
